Print message when starting buckd
Summary:
Buck2 should just work, but sometimes it doesn't, and when it doesn't, these bits help.

I think adding two lines of output won't hurt?

Following diff D53598397 ignores more errors during connect to buckd, and if we silently ignore daemon restart, we may miss cases when buckd restarts when it is not supposed to.

Reviewed By: cjhopman

Differential Revision: D53598090

fbshipit-source-id: 9490410dd41d43cc228a900445728dfb4e91c5cf