Don't check for starlark peak memory when in profiling mode
Summary: When we are in the profiling mode, we are accumulating more memory which might trigger the starlark peak memory check if peak memory goes above 2GB. There was a [recent user report](https://fb.workplace.com/groups/buck2users/permalink/3592263701029985/) that complain about this issue. Let's disable that when in profiling mode so that we can profile without unnecessarily increasing the limit (which was the hack used by the user as a workaround).

Reviewed By: stepancheg

Differential Revision: D53663224

fbshipit-source-id: cb32fc5f6f90609b9bcd6b287e4891e4a0ebd17d