Add lifetime to DiceComputations
Summary:
This will allow us to cheaply create things like what we are doing for
DiceComputationsParallel as we'll be able to have a reference to the underlying
DiceComputations or CoreComputeCtx or whatever rather than needing to clone an
Arc/Weak.

This allows us to really cheaply add and use the BadDice introduced in a later
diff that is the key to incremental migration to the &mut api.

Reviewed By: stepancheg

Differential Revision: D53678525

fbshipit-source-id: 1f48067b39cb84a74652d609d54417bcd5366e2f