Make OpaqueValue no longer capture the DiceComputations
Summary:
This is required for the switch to the &mut api, otherwise once we
have the OpaqueValue we wouldn't be able to call anything else on the
DiceComputations. This is safe as the OpaqueValue itself doesn't provide any
access to things anymore, and instead they need to call functions on the
DiceComputations (which will be &mut) to get the data out of it.

Reviewed By: stepancheg

Differential Revision: D53678529

fbshipit-source-id: d47ac05be0c510df467bd48737c350b921ed0190