Remove HasFileOps
Summary:
Instead, have things use DiceFileOps directly. With HasFileOps, the
pattern is that you get a FileOps instance early on and carry that around, but
if that captures a ref to the DiceComputations, you can't use that
DiceComputations for anything else (once on the &mut api) and that just doesn't
work.

Using DiceFileOps directly, we can pass around the DiceComputations and create
the FileOps only at the point we need it.

Reviewed By: stepancheg

Differential Revision: D53715265

fbshipit-source-id: 055b2cfe32bb6a7fa57f1dba10c088259a3d0fd2