Export higher_order_closure from dice
Summary:
This is needed anytime someone tries to use one of the parallel
compute functions. It's somewhat disruptive to need to update cargo/buck files
for that (both needing to update them but also that you need to do that to get any
useful info out of rust-analyzer and then rust-analyzer needs to reindex
everything).

Easier if we can just get it out of dice that anything using it would already
depend on.

Reviewed By: stepancheg

Differential Revision: D53798488

fbshipit-source-id: 8667934427814abbc011a6c3c0eb7cef7c080e29