Rewrite package listing to not depend on bad_dice()
Summary:
This is an interesting one because it's previous approach would
totally discard everything about the data dependencies that the new dice &mut
api requires to be explicit. That means it's not just a matter of converting a
parallel thing into using compute_many, it requires completely reworking how
the computation is done.

The two options here were to do what I do here, make it directory based and a
big tree computation or to do it layer based (i.e do a compute_many for the
first level dirs, collecting all the next level dirs and then compute_many
those, and repeat). I'm pretty sure one of my early approaches to doing
package listing (when it was first implemented a couple years ago) was more
level based and that was slower than the approach that could traverse them
independently. Other than that, neither seemed particularly simpler than the
other so I went with the tree-based approach.

This approach may be worth revisiting when we better understand the cost of a
high-branching computation vs a simpler one.

Reviewed By: stepancheg, JakobDegen

Differential Revision: D53715261

fbshipit-source-id: b61b8edb0e108fb2625a12f2aabed16286686998