Send the whole artifact tree as remote object to TPX
Summary:
Previosly, we'd sent only the digest of the output, however in buck `ArtifactValue` represents the whole artifact tree and is known to buck.

NOTE: when an artifact tree contains a symlink we fall back to materializing files on disk and local host processing. This is a temporary limitation that is supposed to be lifted soon.

Reviewed By: asm89

Differential Revision: D53763435

fbshipit-source-id: e937199f0886bd9750999ec7b4d9be9930405c38