Avoid materialization for outputs that are in CAS and support remote
Summary:
Now we'll skip materializing outputs that support remote and are in CAS.  For such output buck will return CAS digest to TPX, and it's up to TPX to properly make use of it.

NOTE: currently, 'in CAS' means that a testing action was executed on RE. When we teach buck to upload local testing outputs to CAS, we can simplify this logic.

Reviewed By: loganwendholt

Differential Revision: D53662322

fbshipit-source-id: de4c6d9199c60982cf8e8fab2f7abefe896b148a