Only check cpu_filters if we are sure the APK has native libs
Summary:
If we only have `prebuilt_native_library_dirs`, then we might not actually have any native libs since those dirs can be empty (e.g. `android_prebuilt_aar` creates one regardless of whether there are any native libs in the prebuilt AAR).

If we have no native libs then we don't want to do the check to see whether the native libs are compatible with the device we are installing to.

Therefore, only do that check when we have shared libraries, since in that case we do know that we are going to have native libs in the APK.

Reviewed By: navidqar

Differential Revision: D53951566

fbshipit-source-id: f0fa3b388270e71b9094cba5e50788eb3f2b997d