Make stripping configurable in Android APKs
Summary:
This allows us to build unstripped .apks and have fully symbolicated crashes in logcat without having to symbolicate the crashes manually.

This is mostly a proof of concept for now. I'm happy to do additional cleanup (rename all stripped_* variables, etc) if we decide to go ahead with this.

Note for reviewers: I'm curious if you agree that this is worth doing.

Reviewed By: IanChilds

Differential Revision: D53854006

fbshipit-source-id: 35a92d56bed0bf00c76d4299fe559e4ac0ef8fbc