Added rc.exe and cvtres.exe to the C++ toolchain.
The rc.exe and cvtres.exe are, respectively, called "Windows Resource Compiler" and "Windows Resource to Object Converter".

I made finding rc.exe in vswhere.py optional since it is part of the Windows SDK which is already being handled as optional in that script. Although I'm not sure the optional part is currently working.