Added windows_resource rule to compile .rc files: https://learn.microsoft.com/en-us/windows/win32/menurc/about-resource-files
The "Windows Resource Compiler" compiles .rc files into .res files which the "Windows Resource to Object Converter" then converts into .obj files, to be used by the linker.

The .res files can also be passed directly to the linker which will call the "Windows Resource to Object Converter" in the same directory to convert the file. Since with Buck2 we want explicit dependencies as much as possible I opted to convert the resource file to an .obj file explicitly in the windows_resouce rule.