Create a new module `systemd.rs` and add a gating function `is_available`
Summary:
This is a very first step of resource control implementation. `is_available()` will be used as a gate combined with a `buck2_resource_control.enforce` in `.buckconfig` in order to determine if the daemon should spawn processes via `systemd-run`.

In upcoming diffs, all properties and functions will be placed in this `systemd.rs`.

Reviewed By: JakobDegen

Differential Revision: D54101575

fbshipit-source-id: 424e4e2a2c478f48ce716b5ec89b94bdd8926959