Inline DiceCalculationDelegate::eval_module into InterpreterCalculation::get_loaded_module
Summary:
eval_module has been a weird thing in DiceCalculationDelegate in that
it is actually cached itself on dice. This has caused confusion before where
something used the cached or uncached version of eval_module when they wanted
the other. Previously we'd changed the naming to try to help, but probably moving
it out of DiceCalculationDelegate entirely is even more appropriate. It doesn't
use any of the data from the delegate either.

Reviewed By: JakobDegen

Differential Revision: D54212629

fbshipit-source-id: 2706dbac3bc52384a49c73d97a17c8a57d9c533c