bxl: Fix `Display` on `StarlarkAnalysisResult`
Summary:
The current display implementation forwarded to the debug implementation, which included a whole lot of weird stuff like the deferred table and also the entire target, even if only the subtarget was requested.

That would probably have been pretty surprising to users, but practically it didn't matter because the display impl was too slow to matter on even medium complexity targets (I killed it after like a minute).

Just print the provider collection instead

Reviewed By: cjhopman

Differential Revision: D54046243

fbshipit-source-id: 449d9c20115048e8a3a41f3971847d3b79d9d6c1