buck2 should respect cxx.link_weight in case of fallback
Summary:
The default cxx_toolchain_impl assumed a link_weight of 1. Most
places were already updated to propagate link_weight through from the cxx_toolchain.
For CI though we pass a custom cxx.link_weight on Windows machines to avoid OOMs
(as well as some other platforms), and this was not respected in buck2.

Although unlikely, because it depends on RE falling over causing a thundering
herd of link actions to run locally, we should at least not OOM as a result.

Reviewed By: cjhopman

Differential Revision: D54141738

fbshipit-source-id: 2285bb94f7d6cd5fee9d974d4088a40ad67b79bf