Accept `cells` in place of `repositories`
Summary:
Cells currently have a pretty consistent name. This name is used almost everywhere - except for in the buckconfig sections, which are called `repositories` and `repository_aliases`. That's unfortunate.

Now, we can't fully change that until buck1 is dead, and if this just mattered internally I'd say to wait until then. But it matters for open source users too, and I think we should continue to force the bad naming onto them just because we haven't killed buck1 yet.

So this diff proposes to additionally accept the more sensible names. Maybe one day we'll also stop supporting the current names, or maybe that's not worth the effort. In either case, this will let me consistently use "cells" in the open source world, which is a win.

If we're generally ok with this, I'll send additional diffs to update our OSS examples and `buck2 init` to use it

Reviewed By: lmvasquezg

Differential Revision: D54101998

fbshipit-source-id: 3c84b3a70b9c4b3493f0673fc0651a677a39a848