Define resource control config as a daemon startup config
Summary:
Add 2 configs with regard to resource control, 1) activation switch and 2) max memory limit, to daemon startup config.

The activation switch config (`status`) can take one of `{off | if_available | required}`. In the upcoming code, this will be used to determine if the resource usage of daemon/worker process should be controlled or not. `off` means the resource will not be controlled and buck2 does the stuff in the same way as previous. `if_available` means the resource will be controlled by systemd if it's available on the system, otherwise off. `required` means that the resource will be controlled by systemd, and if systemd is not available on the system, buck2 will error it out and the buck2 invocation will return with an error exit code.

The max memory limit (`memory_max`) is the threshold which the buck2 daemon / worker process is allowed to hold up to. In the upcoming code, it will be set to `systemd`'s `MemoryMax` property. The units like "MB", "GB" can be used, or also "%" can be used in this field. The behavior when the process tries to allocate more memory than this is that OOMKiller will kill it. There are many other properties that systemd can control, but for now, we start with only the memory max as a configurable property as most people desire.

Having these configs in daemon startup config will trigger a daemon restart when any of the configs are changed, which is intended and ideal.

Reviewed By: scottcao

Differential Revision: D54210376

fbshipit-source-id: 4534447e55b9157dcfe3437d37535814ee57d975