Add AttrFmtOptions
Summary:
We need to have several different ways of formatting attrs. Currently
we are abusing the alternate format just to signal that we want strings to be
printed without quotes. Instead, we want to align with how starlark values use
the alternate flag, and so we need a different mechanism to get the current
quote-less behavior.

Reviewed By: iguridi

Differential Revision: D52496169

fbshipit-source-id: e85ea7534090499a8dcc2a177a33a94a312d07b5