Add DiceFileComputations
Summary:
This is an alternative to DiceFileOps. DiceFileOps needs to fit into
the FileOps api, which adds restrictions that aren't generally necessary for
things interacting with files on the dice graph.

For places that don't specifically need to use a FileOps, DiceFileComputations
allows them to access files without those restrictions (specifically, FileOps
is going to use a LinearRecomputeDiceComputations because the FileOps api is
non-mut based).

Reviewed By: JakobDegen

Differential Revision: D54227584

fbshipit-source-id: 3819044fc1dde4057c4351c08d3d4e2893b9861f