Introduce DiceComputations::with_linear_recompute()
Summary:
There's a couple final places where we are using a &DiceComputations
that are (1) non-trivial to rewrite to use &mut DiceComputations and (2) are either
never recomputed or only recomputed when invoked from bxl.

Rather than need to fully migrate them to the &mut api, introduce an escape hatch that
vends out owned DiceComputations but will flatten the deps into a linear list
for recomputation.

Reviewed By: JakobDegen

Differential Revision: D54227590

fbshipit-source-id: 6f5fcab8095e5d62e0372daed66eae9efc70cc42