Introduce BuckConfigsViewForStarlark
Summary:
This simple trait just provides a get() for two configs (current_cell
and root cell). For &mut DiceComputations, buckconfigs in starlark pose a minor
problem: we want to support them both for dice-based on non-dice-based (for
testing) configs, but the dice-based ones need to borrow the &mut
DiceComputations and we want to keep that contained within there (because (1)
we don't have it in the non-dice-based case and (2) where would we put it), but
then we can't hold two separate instances of it like we currently do. So, this
trait will allow us to have a single borrow that is used for both cell configs.

Reviewed By: JakobDegen

Differential Revision: D54221264

fbshipit-source-id: 2c4171bad8b352d94287935a9adb31e32407031d