Remove "build stream" related bad_dice
Summary:
Converts this all to be based on LinearRecomputeDiceComputations. The
main thing using this is just related to the top-level of buck
commands which don't do recompute anyway. This also affects bxl's build(), but
even there it's limited to just linearize the work for a single target which should
be immaterial.

Reviewed By: JakobDegen

Differential Revision: D54227583

fbshipit-source-id: 6cecc5f5855bd32f24462a6f893423c09a2c479d