Fast provisioning profile reading
Summary:
We currently spin up a shell and run openssl on each provisioning profile in order to extract the plist content from a provisioning profile, which is quite slow.

In practice, we can just extract the bytes that occur between `<plist` and `</plist>` and achieve the same thing much faster. This isn't the most elegant solution, but by avoiding spinning up ~400+ shells invoking openssl on each profile, we can see meaningful performance improvements in codesigning.

Differential Revision: D54390550

fbshipit-source-id: 0facf313a78958d931b2eb87b3ec08386016aad1