fs utils: `IoError::categorize_for_source_file`
Summary:
This is an interesting case, and shows that "categorize from one tag" doesn't really work. Both `IO_NOT_FOUND` and `IO_SOURCE` are not independently sufficient to categorize the error, but when they appear together they are.

It's not quite clear to me how we should best handle this case. Maybe we should build out the categorization logic in `buck2_error` to support it, but that has some disadvantages: First, it's not clear how that API should look, and second, that risks separating the logic across the codebase in a pretty awkward way. We could imagine adding a separate tag for this case, but that doesn't seem great either.

For now, just use the old categorization, and then think more about this.

Reviewed By: stepancheg

Differential Revision: D54460214

fbshipit-source-id: c51dd74e6b86c363e200f8aaf59cf5845485f808