Allow jars with native libs to not have an inner jar
Summary:
Our spark pipeline uses JNI and requires native libs. Per my discussion with sfilipco mtrepanier an inner.jar won't work for this purpose.

This diff essentially reverts D46366872 and removes a safeguard that prevents users from creating a java binary with a main class that does not have an inner jar but does have native libs. This could result in adopters misusing this feature for environments where an inner jar is supported.

Alternatively, we could alter wrapper that builds each java binary separately and includes it in the class path.

Per sfilipco
"The wrapper mode in java_binary is **not portable** in it’s current format.  The wrapper is just a series of symlinks. The symlinks that buck builds today are absolute paths. Only when a full wrapper build locally you can guarantee that these absolute paths exist on the local machine.
"I have actually seen this being an issue in regular development. When we did clean build on master, due to remote execution (+sandcastle) + artifact caching, the master builds downloads a cached wrapper which point to the absolute paths from a Sandcastle machine but the wrapper is not a cacheable artifact (I dunno if wrapper builds were marked as not cacheable at this point). If you make any changes locally and build then buck knows to rebuild the wrapper and you can run your application.
"It may be possible to build relative symlinks but “packaging “of the final “artifact” has to be done before the symlinks are build. Theoretically possible but more work than what we have now.
"Second point is about native builds. To my knowledge the fundamental idea of the inner.jar is a means towards managing all the different kinds of paths that each piece of code expects and its goal is to make the application portable. Either way, you have to figure out a way to manage all the these paths that an application has.
"I use portable to mean that you can move between different machines. Not that the same artifact runs on both Mac and Win.
"Hmm, then there's definitely an opportunity to either modify wrapper or add another param like ensemble or jars that instead builds each java_binary separately and includes them on the classpath. Hell, they could even just stuff them in lib as unmanaged jars"

Reviewed By: IanChilds

Differential Revision: D54217079

fbshipit-source-id: 9b36e1a9956a20d21bd2ff849187c9d6aa3c544f