Add --all-cells
Summary:
Lets you print out config for all the cells in one go. This is super helpful if you want to configure a system that needs to know about a lot of Buck settings (e.g. BTD).

Only real choice was for the JSON - you could either have a dict of dict, or use the `cell//section.key` syntax. I went for the latter, since Buck1 already mandated that we can put such syntax in the JSON, so rather than have two ways just be consistent.

Reviewed By: JakobDegen

Differential Revision: D54535728

fbshipit-source-id: f4885f2ab139f74a7c67ff89904f08658ae27429