Change SystemdRunner interface
Summary:
Previously, `SystemdRunner` struct was just there to store the settings like properties, and `background_command_linux()` was to return the `Command` based on the information. This produces little benefit as being a struct, and furthermore, when we consider to reuse this code for both daemon and forkserver in the future, the struct's interface should be corrected for better use.

For the enforcement config & availability check, the logic is identical for both cases of daemon and forkserver, so now the code is unified as `SystemdRunner::create_if_enabled()` to be able to be called from both.

As for systemd's command line arguments and property settings, the daemon and forkserver will have almost the same set of arguments, so now `SystemdRunner::create()` (non-pub) provides the way to switch the argument-set based on the case `SystemdPropertySetType`.

The unit name and working directory should be changeable between the `background_command_linux()` calls because forkserver will spawn each worker with different name and cwd, so now the function `background_command_linux()` takes those information at call time.

There is no meaningful ways in unit tests to test those new functions because `is_available()` is incorporated into them, which is platform dependent. The previous unit tests are deleted by this diff and E2E tests will cover the lifecycle of `SystemdRunner` in next upcoming diffs.

Reviewed By: JakobDegen

Differential Revision: D54382039

fbshipit-source-id: 1a613c0e5815d9fc042ab770ed3b4f0c9c04264e