cells: Move `buildfiles` to file ops
Summary:
Consider the case for external http cells: We could try and keep the process of downloading them outside of dice. However, there are some obvious disadvantages to that, namely that it means we probably have to do at least a `HEAD` request for every external cell on every command. It also doesn't vibe with the long-term plans for the codebase, having more stuff outside of dice is generally going to make us sad, not happy.

So instead, let's try and do external cells in a way that supports lazyness; the primary barrier for implementing that is that we currently have an injected `CellResolver`, which provides dice-free access to `CellInstances`, and those `CellInstances` provide access to the following two problematic components: `buildfiles` and `cell_alias_resolver`. Getting access to those two fundamentally requires parsing the relevant cell's buckconfigs, which means that we'd have to download the external cell.

As such, we need to move `buildfile` and `cell_alias_resolver` access behind dice. Let's begin with the buildfiles. Writing this summary after having written all the code, I can confirm that this didn't turn out to be too hard and involved basically no compromises.

This diff is the only sort-of compromise that we have to make. The package root listing code needs to know the buildfiles, and it only has access to a `FileOps`, not dice. Changing that is pretty difficult, especially given the way our unittests are written. As such, let's avoid the issue by putting buildfile access onto the `FileOps` trait. It's kind of a hack, but probably the least bad of all worlds

Reviewed By: stepancheg

Differential Revision: D54663715

fbshipit-source-id: 903c0aad224ce40116869776547995b6395a61a8