cells: `CellResolver::get_cwd_cell_alias_resolver`
Summary:
With the buildfile handling moved out of the cell resolver, it's time to turn our attention to the `CellAliasResolver`. Putting access behind dice turns out to mostly be possible and straightforward, but with one big caveat: Handling `//`-style modefiles and config files fundamentally requires access to a cell alias resolver, both in the client and in the pre-dice daemon. Any alternative is likely to require semantic changes of some kind.

As a result, this diff introduces a big carveout in the dice-for-cell-alias-resolver requirement; getting the cell alias resolver of the cwd will not require dice.

Specifically for the purpose of external cells, this is good enough, since your cwd can't be inside an external cell. For a broader buckconfig-on-dice goal however, it may not be good enough in the long term.

Reviewed By: iguridi

Differential Revision: D54663707

fbshipit-source-id: 9f8fc6e518f099fea93050939f0bbb12e3f8cbdf