improve BuildArtifactMissing error message
Summary:
The debug output printing here is just ridiculously unreadable. Change
it to just print the paths of artifacts. We don't expect that this is ever hit
in normal operation (a bug fixed later in this stack causes us to hit it now), so
we don't need to be too picky about the message being just right.

Reviewed By: JakobDegen

Differential Revision: D54437210

fbshipit-source-id: fd52318782e248dad2adb84d5b04af54343071d2