Fix Hash/Eq/PartialEq implementation for BuildArtifact
Summary:
A good approach to thinking about the correctness of a
Hash/Eq/PartialEq implementation is to consider whether it would be safe
to intern instances of an object based on it. It would be rare that we would
want to make that unsafe.

ModernDice uses the expectation that and dice key's hash/eq/partialeq are
trustworthy enough to use it for interning, for example.

Hiding these fields clearly violates that expectation.

Also, I can find no code that depends on the behavior of these being hidden.
Previously, actionoutputs lookup was artifact-based, now it's path based.
StarlarkArtifact already uses its own fingerprint() for hash/eq and is
unaffected by this.

Reviewed By: JakobDegen

Differential Revision: D54437208

fbshipit-source-id: cd9aab9392290367dda54625d63fc5d6f685d4c8