Collapse use of ::types:: (A-B)
Summary:
This diff is part of the codemod for https://fb.workplace.com/groups/rust.language/posts/25780668961555042.

Thrift-generated data structures, clients, and service traits all used to be exported in a single crate, with the data structures being under a `types` module. Now that data structures have a crate of their own, the `types` module is vestigial: `foo::types` refers to the same contents as `foo`.

Generated using the code in P1195926494.

1. Run `buck2 bxl fbcode//common/rust/tools/scripts/rust_targets.bxl:rust_targets -- --simple=true > /tmp/targets` and pick an appropriately sized subset of targets to work on.
2. Add a `#[deprecated = "…"]` attribute on items in the old module structure (D54872084).
3. Build all the targets from step 1 and collect their compiler output using the diag.json subtarget.
4. Parse diag.json and apply the required changes to the Rust sources.
5. `arc f`

Reviewed By: zertosh

Differential Revision: D54838520

fbshipit-source-id: f6db5f08932c1e3ec9509ec4766c2eab1f601ba4