Collapse use of ::types:: (misc)
Summary:
This diff is part of the codemod for https://fb.workplace.com/groups/rust.language/posts/25780668961555042.

Thrift-generated data structures, clients, and service traits all used to be exported in a single crate, with the data structures being under a `types` module. Now that data structures have a crate of their own, the `types` module is vestigial: `foo::types` refers to the same contents as `foo`.

This diff includes files that slipped through the cracks for various reasons in the preceding codemods.

- Macro-generated code
- Weird re-export situations
- Doc comments
- Handlebars templates (privacylib)
- New code added within the past day

Reviewed By: emersonford

Differential Revision: D54887805

fbshipit-source-id: e75ac93167ad4ab5d0e5ac8345932a43e079638c