Remove target configuration from ClientContext
Summary:
`ClientContext` is contained in every request

https://www.internalfb.com/code/fbsource/[686d4920546ac8e9c4bcf2b8548bb30397aad087]/fbcode/buck2/app/buck2_cli_proto/daemon.proto?lines=91%2C95%2C130

but not every request should have `--target-platforms=`. For example, `uquery` should not need it.

This changes protobuf, but does not change commands (yet).

This diff moves two fields mentioned above into a separate message `TargetCfg` which is added to all requests where it is used now.

Reviewed By: JakobDegen

Differential Revision: D54928133

fbshipit-source-id: 043e033aa29cd44b8ce1a0af50fa6b6c566bc48d