rules: Separate fast metadata graph
Summary:
Before this diff, fast metadata generation still used full metadata from its dependencies. This diff changes that, and effectively puts fast metadata onto a completely disjoint artifact graph from the remaining builds.

For most purposes this is undesirable because it prevents sharing. However, in some cases that doesn't matter. Specifically, I have in mind rust-analyzer check builds, which are anyway on a separate isolation dir that will never see full builds.

I'm not planning to expose this as a subtarget. We can add that later if we want to, but the problem is that we probably wouldn't just need `check-fast`, but also `diag.json-fast`, `clippy.json-fast`, etc. and that gets very messy very quickly. The plan for now is to just access this thing via BXL, more details in coming diffs

Reviewed By: dtolnay

Differential Revision: D54977133

fbshipit-source-id: 81f6f40e489c214fbce08c2fe1cbced3b65008fe