Assert dynamic_output lambda returns None
Summary:
The idea is this: documentation is bad. So many people write code by trial and error. By adding such assertions we make trial and error cycle faster.

For example, it is not obvious that dynamic output lambda is not meant to return providers. This changes tells it to a user.

Reviewed By: JakobDegen

Differential Revision: D55271890

fbshipit-source-id: 0737ce2599edb7aecba0670c18710e9d6ae09523