Refactor derive
Summary:
Introduce functions like:

```
impl DeriveInputUtil<'a> {
    /// Generate expressions with callback for a struct or for each enum variant.
    pub(crate) fn match_self(
        self,
        handler: impl Fn(StructOrEnumVariant, Vec<(syn::Ident, &'a Field)>) -> syn::Result<syn::Expr>,
    ) -> syn::Result<syn::Expr> { ... }
}
```

to avoid repeating the same code of unpacking structs/enum and patching back over and over.

This diff was written to fix `#[freeze(identity)]` for enum variants. It is fixed now because struct and enum fields are now handled uniformly. Test for it is added in the following diff D55289200.

Reviewed By: iguridi

Differential Revision: D55289166

fbshipit-source-id: 0b856f997fafd7735d89fe58c18310adbf177330