clean-stale: track sizes per path
Summary:
Instead of keeping track of stats and paths separately while scanning buck-out, track the sizes of each artifact found so that if clean is interrupted, we can report how many bytes were actually cleaned.

This also simplifies the logic in `visit_recursively`

Removing `cleaned_path_count` because it is equivalent to `cleaned_artifact_count` since this was changed to always clean files individually.

Reviewed By: iguridi

Differential Revision: D48783449

fbshipit-source-id: aca2c6905e5dd0638db0f94bf28391269ffec986