Add LivelinessObserverSync
Summary:
Allows using LivelinessObserver in both sync and async contexts which is needed for clean stale, used in following diffs.

This method could be added to `LivelinessObserver` instead of introducing a new trait if it's needed, but it's non-trivial to implement and test on some of the types that already implement `LivelinessObserver`, specifically `CancellationObserver`.

Reviewed By: JakobDegen

Differential Revision: D49039584

fbshipit-source-id: 68d7cdda29897e0afbb8b81069b1e5643cd54240