clean-stale: interrupt clean if another command starts, instead of blocking
Summary:
The current strategy of running clean stale when the machine is idle (or very likely to be idle) won't work well on laptops or desktops. By making it interruptible it can be scheduled whenever and make incremental progress without interrupting or blocking users.

This uses a `LivelinessGuard` that is checked continuously during the clean and stops any expensive work (scanning or cleaning) as soon as it's dropped.
The `LivelinessGuard` is held in the `MaterializerSender` because it should be dropped by any commands sent to the materializer command processor (or else the command will block). The guard can't be inside the command processor run loop or command stream because that runs on a single thread.

This doesn't interrupt on `low_priority` commands because they shouldn't block user commands although it shouldn't make much difference one way or another.

Also removed the "exclusive command" lock which just exists to show a message to the user if a build is blocked by clean, some code for that still exists in case it's useful for something later but it could be removed if needed.

Reviewed By: scottcao

Differential Revision: D48783448

fbshipit-source-id: b278fda3589c3dce1d53b943ab2305874cde8d62