Add extra_cfg_modifiers_per_rule parameter to set_cfg_modifiers function
Summary: This diff adds `extra_cfg_modifiers_per_rule` parameter for setting modifiers targeting specific rule names in `set_cfg_modifiers` function. This parameter is a dictionary where the key is the rule name to target and value is a list of modifiers. Added some documentation in docstring to explain what this parameter does. This diff just plumbs in this parameter in the frontend but does not use this parameter any differently than normal modifiers at the moment. Next diff will use this parameter to actually target PACKAGE modifiers to specific rules.

Reviewed By: cxxxs

Differential Revision: D54167840

fbshipit-source-id: 4bea4cc3399da4abf5e5354263c33ef50792d718