Define the apple_native layer
Summary:
# This Stack
Right now, our platform configuration layer is baked deeply into the FB macro layer. We're not ready to move to configuration modifiers yet, and for use in WhatsApp/OpenSource, or even our basic e2e tests, we need better ability to configure for different sdk/arch/build_mode combinations.

We also want a central point that when we move to modifiers, its clearer what can be cleaned up, and integrations can be better designed. As such, we're going to introduce an `apple_native` layer, which ensures that `apple_*` targets have appropriate default target platforms and `_apple_platforms` set for build, and discovery in project generation.

# This diff
We now have everything we need to define the `apple_native` layer which sets the configuration we need on apple targets.

Reviewed By: blackm00n, rmaz

Differential Revision: D55445133

fbshipit-source-id: fb8321e9eb848ac24b69e32e929c2ae495f1b0a4