Extract the default target mapping used in fbsource
Summary:
# This Stack
Right now, our platform configuration layer is baked deeply into the FB macro layer. We're not ready to move to configuration modifiers yet, and for use in WhatsApp/OpenSource, or even our basic e2e tests, we need better ability to configure for different sdk/arch/build_mode combinations.

We also want a central point that when we move to modifiers, its clearer what can be cleaned up, and integrations can be better designed. As such, we're going to introduce an `apple_native` layer, which ensures that `apple_*` targets have appropriate default target platforms and `_apple_platforms` set for build, and discovery in project generation.

# This diff
As this diff says, when dealing with setting defaults, we need the default mapping. To ensure one source of truth, extract this into prelude, and make it available in fbsource.

Reviewed By: rmaz

Differential Revision: D55445131

fbshipit-source-id: 4308116bb9ce8993c591ded6f3cd9e82417897aa