Create linker wrapper to diagnose unexported symbol problems
Summary:
A developer may make a change that introduces a cross dylib symbol
reference that didn't previously exist. This means a symbol that
was previously not exported, must now be exported.

This linker wrapper intercepts linker error messages and compares
them against the current set of unexported_symbols lists to diagnose
such errors, and update the lists by invoking a tool.

Reviewed By: narissiam, rmaz

Differential Revision: D55246335

fbshipit-source-id: db0aaa5e2eeb1d6db28594be3f3abad29c0abacf