Track scan and clean time for clean stale
Summary:
Need some way to track clean duration which we don't have if this isn't logged as a buck command.
It's also useful to know how much time is spent in two "phases", scanning buck-out and cleaning artifacts.

Tracking `total_time_s` separately since we won't have scan or clean duration if we hit an error and it may still be useful to know if this runs for a long time before failing.
Ideally hitting an error wouldn't require constructing a new CleanStaleStats to hold this, but it doesn't seem critical to have more detailed stats if there is an error.

Reviewed By: iguridi

Differential Revision: D55533257

fbshipit-source-id: d09d6723910047362200064474c55196403dbcd5