Re-land linker wrapper without needing to select on linker
Summary:
D55246335 was reverted as it passed flags intended only for lld to other linkers
such as ld64.

Apparently selecting on the linker is difficult, so the wrapper script is now passed the same arguments for all linkers, and instead discovers the selected linker using the -fuse-ld= flag, and modifies flags forwarded to the linker accordingly.

Reviewed By: rmaz

Differential Revision: D55640540

fbshipit-source-id: 54711aeb5e80b0b2a77c31b646032c1ea4a33d1a