file watcher: Remove `ignore_specs` handling
Summary:
Currently, the buck2 file watchers are implemented to be aware of the ignore specs, and avoid sending file changes to dice if the changed files are in the ignore spec. This diff removes that logic.

## Why

 1. The implementation of the current behavior is wrong. The ignore specs (and more generally, the cell resolver) are populated once at daemon startup and then never updated again. See D55573384 for the full details of this bug.
 2. This bug is extremely easy to hit in a common external cells workflow (specifically, `buck2 expand-external-cell`)
 3. Fixing the current implementation is incompatible with moving buckconfigs into dice. If buckconfigs are only calculated in dice, the ignore specs, which are calculated from the buckconfigss, are not available at transaction update time.

## Performance

I expect this to be mostly performance neutral. While this will result in additional updates being sent to dice, the keys being updated should have no rdeps and so these updates should be cheap. If they do have rdeps, then there's another bug since we weren't getting updates for these keys.

The file watcher watching and receiving updates for directories it does not need to can be expensive. However, the ignore specs were not used for filtering out watcher queries, and so this diff should have no effect on that; on the other hand, it is possible that we will now start seeing "file changed" messages in console in cases where we weren't before. This is good, since it will tell us about any additional directories that need to be excluded in the `.watchmanconfig`.

Reviewed By: stepancheg

Differential Revision: D55573383

fbshipit-source-id: 29e843999c297de356ac8f578927971acd2015f7