Fix jacoco coverage wrapper failing due to duplicate classes
Summary:
The underlying problem here is that there is a duplicated androidx.databinding.* class in jars fed to Jacoco and it fails on report generation. Those are generated at compile time and seem to be embedded in the jars throughout we have no way of knowing this due to how we discover classes for the purpose of Jacoco report

Excludes don't work at agent level:
https://github.com/jacoco/jacoco/issues/1093

So a coverage wrapper needs to know in advance classes and exlude them specifically. Passing a classname even without associated source file will allow us to do so

This should land only after a previous diff is rolled out

Reviewed By: asm89

Differential Revision: D55454598

fbshipit-source-id: 7624335110356c25415d462bde53e158e152dd4a