Calculate buck root relative to an explicit path
Summary:
Previously, we always calculated the buck project root relative to the current working directory. This is wrong when the user has multiple repositories open, as we'd always use the first repository.

Instead, use an explicit path. When we check a file, ensure we use the buck root for that file specifically.

Reviewed By: davidbarsky

Differential Revision: D55763716

fbshipit-source-id: dfca2e3b921d7654717990b6b4beecb804945bc6