prelude: Remove some repository decls
Summary:
Instead of declaring these explicitly, let's just use the ones we "inherit" from the repo root. This doesn't really work properly in buck1, but in the prelude we don't care.

The benefit of this change is that 1) it reduces differences between OSS and internal and 2) it means that we can use the prelude as an external cell in an isolated test without changes

Reviewed By: stepancheg

Differential Revision: D55783828

fbshipit-source-id: 7579baa42d0a978c679865646cb6b112db3baac7