file watcher: Remove a panic
Summary:
This diff is a reland of version 1 of D55573383.

Stiopa asked in that diff why we should ignore this instead of returning an error.

In all cases, the behavior of this code as it exists today is incorrect, specifically because it results in only partially picking up changes to the ignores list after daemon startup. The incorrect behavior that I've opted for in this diff, however, is the more conservative one. If it knows that its ignore lists have gone out of sync with reality, it chooses to treat the file as not ignored, which may lead to unnecessary dice invalidations but is otherwise always correct. Returning an error is less correct, since the error would go away on a daemon restart.

Reviewed By: stepancheg

Differential Revision: D55830767

fbshipit-source-id: f25382dbbb76948b3aca38c4a7415f4060a49e06