trait DeferredOutput
Summary:
`AnyValue` (which is `Any` for non-trivial deferred output) is implemented for any `T`, which makes code navigations somewhat not easy.

https://www.internalfb.com/code/fbsource/[f74ed5fa3f6948d77dcda6a8dc30a51ab0ff80b0]/fbcode/buck2/app/buck2_build_api/src/deferred/types.rs?lines=644%2C650-653

In particular, when looking at structs like `DeferredValueAnyReady` it is not clear, what is inside of `AnyValue`:

https://www.internalfb.com/code/fbsource/[f74ed5fa3f6948d77dcda6a8dc30a51ab0ff80b0]/fbcode/buck2/app/buck2_build_api/src/deferred/types.rs?lines=556-562

https://www.internalfb.com/code/fbsource/[f74ed5fa3f6948d77dcda6a8dc30a51ab0ff80b0]/fbcode/buck2/app/buck2_build_api/src/deferred/types.rs?lines=633-636%2C641

This diff introduces `DeferredOutput` marker trait to make all of it more clear.

So now one can do this:

 {F1479781414}

This diff also fixes (kind of) one misleading test which attempts to put `DeferredTransitiveSetData` into `DeferredValueAnyReady`: this does not happen in real code, because `DeferredTransitiveSetData` is `TriviallyDeferred` and does not implement `Deferred` (which does not have `Output`).

Reviewed By: JakobDegen

Differential Revision: D55831138

fbshipit-source-id: 702dfb92731eb42a51a2ad5cf81b41d2964fb87a