configs: Don't ignore other configs if `.buckconfig` is empty
Summary:
Currently, if your cell doesn't have a `.buckconfig`, that's permitted, and results in all other buckconfigs not being applied either (`/etc/buckconfig`, `-c`). That seems pretty unintuitive to me.

Yipu is the one who added this code and said on Discord that he think's it's a bug. Given that all our cells have `.buckconfig`s and so we don't rely on that behavior, let's just fix it

Reviewed By: stepancheg

Differential Revision: D55791966

fbshipit-source-id: b5f52a4f230c6b96df14cf5df38f5f3b081eab7d