Rename `repositories` to `cells` everywhere
Summary:
I landed support for this in D54101998 . It's been a while since then, so seems appropriate to update our docs, examples, etc. with this as the new default. This diff is mostly `sed` powered. I've left out e2e tests for now.

The most important change this is making is to `buck2 init`, since this way we stop showing the bad name to OSS users.

Reviewed By: stepancheg

Differential Revision: D55791938

fbshipit-source-id: 455bcb7a1a3c6b032dccc516fc4718d9d1254aff