file ops: Move ignored handling out of `FileOpsDelegate`
Summary:
Instead of requiring the `FileOpsDelegate` implementation to handle ignored paths, make a wrapper type `FileOpsDelegateWithIgnored` that accepts an opaque `FileOpsDelegate` and is responsible for dealing with the ignored handling.

This means that we won't have to duplicate the ignored handling across multiple implementations of `FileOpsDelegate`.

The `AllCellsIgnored` field is made optional. In this diff, that's only done to support testing, but the next diff will use it in non-testing codepaths.

Kind of a big diff, but mostly pretty mechanical transformation

Reviewed By: stepancheg

Differential Revision: D55791940

fbshipit-source-id: 1638034325a7f7e4830f3fd8556421ad5c827e47