cells: Add `DiceFileComputations::read_dir_no_ignores`
Summary:
Currently, if we tried to do buckconfig parsing via `DiceFileComputations`, we would always create a cycle through the ignores checking - the parsing will depend on the `FileOpsDelegateWithIgnroes` which will depend on the `AllCellIgnores` which depends on the parsing.

In order to make this possible, we'll add a `no_ignores` variant of `read_dir`. Additionally, we update all the other code paths to make sure they don't depend on the `ignores` if they don't need to.

I had considered other alternatives to implementing this. Unfortunately, none of them are super pretty, and they basically all require increasing the number of dice keys. Suggestions for better implementations welcome.

Reviewed By: stepancheg

Differential Revision: D55791950

fbshipit-source-id: e11433c6932f1e5f97f2df831f59c0737e720505