configs: Implement dice backed config parsing
Summary:
This is how we'll do config parsing for external cells. It also lays the groundwork for what a more complete "in-dice config parsing" could look like in the future.

This function cannot be called just yet, because there's no way to get the resolved command args. Next diff will deal with that by inserting them into dice, after which we can get to the meet of the external cells implementation

Reviewed By: stepancheg

Differential Revision: D55791959

fbshipit-source-id: e58a160b497939bea6c9e8a6e4961384cc6d2900