Create a dispatcher for events not associated with a command
Summary:
The dispatcher returned by `buck2_events::dispatch::get_dispatcher()` requires that a trace id be set, and is explicitly not intended to be used from threads that are not tied to a specific command, otherwise it returns `EventDispatcher::null()`.

Clean stale, when triggered by the materializer command loop, is an instance where we need to send events to scribe that are not associated with a specific command. For this we can use a dispatcher that is only associated with a daemon and that only writes to scribe.

Reviewed By: iguridi

Differential Revision: D55760394

fbshipit-source-id: 3ee941f45e53f42e592e36a7003574fbb12d09a2