artifact: Support projecting source artifacts and artifacts declared in other targets
Summary:
D37034075 added a requirement that only artifacts declared in the same target could be projected. The discussion for this was in D36865880 . The reason given was that 1) this leads to better error messages, and 2) Neil thought that this was conceptually right.

For the better error messages, those were never actually implemented and that task remains open: T122966509 . Even if we did want to implement it though, there is nothing stopping us from still doing that in those cases where the artifact is declared in the same target.

Chris also had this to say when I asked him:

> the only motivation seems to be so that we get error messages early. I do think that's good. between the two options of (1) project artifact in rule that makes it and (2) project artifact in consuming rule, I would prefer (1).

> But maybe the real choice is that you want to consume some file within a directory artifact and you can't do (1) and so you can either (2) project artifact in consuming rule or (3) consume the directory and then use the artifact in it. In either of (2) or (3) you're going to get a "late" error and the (3) error is probably significantly worse than the (2) error (because writing the action you probably aren't even thinking about making it a good error)

I agree.

The additional reason given for this restriction was:

> I'm not sure you should be allowed to create a projection artifact in another rule. The artifact you get given by a rule should be available as an indivisible lump, and only it should get the option to project it.

I disagree. Rules almost *always* make assumptions about the artifacts they consume. They assume the artifact is a directory, assume it's a file, assume it's an executable, assume it's an object file that provides a particular symbol necessary for the linker invocation to succeed, etc. I don't see what we get by forcing rule authors to encode those assumptions directly into the action definition.

For those reasons, and because the only person present for the discussion at the time now disagrees with the outcome, I think we should revert D37034075 and allow this pattern. That's what this diff does

Reviewed By: cjhopman

Differential Revision: D55830145

fbshipit-source-id: 820e3d87dbfe56a01888308fda69e77ad9bd488c