Allow scheduling clean stale from materializer
Summary:
Integrating this into the materializer command loop so that external scheduling is not required and this can be run easily on any platform, especially developer machines.

If enabled this schedules a first clean once `start_offset` has elapsed after the daemon is created, and re-runs it continuously every `clean_period` to clean artifacts older than time of clean - `artifact_ttl`.

The expectation is that initially, `start_offset` will be 12 hours, `clean_period` will be 24 hours and `artifact_ttl` will be 7 days.

Right now on devservers cleans are scheduled at a fixed time, but this isn't possible on developer machines, if they are scheduled at night they will likely run when the laptop wakes up in the morning anyway.

If needed we can detect if a clean is interrupted and re-schedule or shift the schedule but it seems reasonable to start with this simpler method first.

Reviewed By: scottcao

Differential Revision: D55511293

fbshipit-source-id: 9684c8249b3cf20e243e1a4e34b44558e2169195