Remove plugins and attrs from ctx while running BXL analysis
Summary:
Return error on `ctx.attrs` and `ctx.plugins` when running BXL.

Before this diff they were available empty.

This is done mostly to not support `ctx.attrs` and `ctx.plugins` when running `dynamic_action` D56092787 in the following diff.

I can revert changes for BXL, but I believe this is tech debt.

Reviewed By: JakobDegen

Differential Revision: D56091933

fbshipit-source-id: 7eb21a6556c2970f1aacf2654b5b6ea3b2c30dfa