Fix select refinement
Summary:
I'm not 100% sure the fix is correct, but at least it does not make it worse: code that worked before, keeps working.

I looked at buck1 implementation, and tried to do something similar.

https://www.internalfb.com/code/fbsource/[f6ed8128ea9de703b7d0856a7977103923df343a]/xplat/build_infra/buck_client/src/com/facebook/buck/core/select/SelectorResolved.java?lines=53-56

I suspect it can be implemented more efficiently, but until we see it in profiler, maybe don't bother.

Reviewed By: JakobDegen

Differential Revision: D56122257

fbshipit-source-id: 3959605cc3bf853af5e25747f5f0e7a8df7ac8e1